Change free functions to be methods
