+++ crates/forge-runner/src/running.rs
@@ -51,7 +51,6 @@ mod execution;
 mod hints;
 mod setup;
 mod syscall_handler;
-pub mod with_config;
 
 use crate::debugging::build_debugging_trace;
 pub use hints::hints_to_params;
